make `paths.appPublic` optional

+++ packages/razzle/scripts/build.js
@@ -167,10 +167,12 @@ ${razzle.port !== '3000' && `PORT=${razzle.port}`}
 
 // Helper function to copy public directory to build/public
 function copyPublicFolder() {
-  fs.copySync(paths.appPublic, paths.appBuildPublic, {
-    dereference: true,
-    filter: file => file !== paths.appHtml,
-  });
+  if (fs.existsSync(paths.appPublic)) {
+    fs.copySync(paths.appPublic, paths.appBuildPublic, {
+      dereference: true,
+      filter: file => file !== paths.appHtml,
+    });
+  }
 }
 
 // Wrap webpack compile in a try catch.
